+++ grader/checks/requirements_check.py
@@ -6,8 +6,11 @@ It checks if requirements.txt exists in the project root.
 import logging
 import os
 
+from pathlib import Path
+
 from grader.checks.abstract_check import ScoredCheck, ScoredCheckResult
 from grader.utils.constants import REQUIREMENTS_FILENAME, PYPROJECT_FILENAME
+from grader.utils.virtual_environment import VirtualEnvironment, VirtualEnvironmentError
 from typing import Optional
 
 logger = logging.getLogger("grader")
@@ -24,12 +27,14 @@ class RequirementsCheck(ScoredCheck):
         project_root: str,
         max_points: int,
         is_venv_required: bool,
+        is_checking_install: bool = False,
         env_vars: Optional[dict[str, str]] = None,
     ):
         super().__init__(name, max_points, project_root, is_venv_required, env_vars)
 
         self.__requirements_path = os.path.join(self._project_root, REQUIREMENTS_FILENAME)
         self.__pyproject_path = os.path.join(self._project_root, PYPROJECT_FILENAME)
+        self.__is_checking_install = is_checking_install
 
     def run(self) -> ScoredCheckResult:
         """
@@ -40,10 +45,25 @@ class RequirementsCheck(ScoredCheck):
         """
         self._pre_run()
 
-        files_to_search = [self.__requirements_path, self.__pyproject_path]
+        requirements = Path(self.__requirements_path)
+        pyproject = Path(self.__pyproject_path)
 
-        is_one_of_files_present = any(os.path.exists(file_path) for file_path in files_to_search)
+        files_to_search = [requirements, pyproject]
+
+        is_one_of_files_present = any(file_path.exists() for file_path in files_to_search)
 
         score = int(is_one_of_files_present) * self.max_points
 
-        return ScoredCheckResult(self.name, score, self.max_points)
+        info = ""
+        if not is_one_of_files_present:
+            info = "requirements.txt or pyproject.toml not found"
+
+        if self.__is_checking_install and is_one_of_files_present:
+            try:
+                with VirtualEnvironment(self._project_root, is_keeping_existing_venv=True):
+                    # Context manager automatically handles setup() and teardown()
+                    pass
+            except VirtualEnvironmentError as e:
+                return ScoredCheckResult(self.name, 0, "", str(e), self.max_points)
+
+        return ScoredCheckResult(self.name, score, info, "", self.max_points)