+++ grader/grader.py
@@ -43,6 +43,7 @@ class Grader:
         self.__is_skipping_venv_creation = is_skipping_venv_creation
         try:
             self.__config = load_config(config_path)
+            self.__logger.debug(f"Config contents: {self.__config}")
         except InvalidConfigError as exc:
             self.__logger.error("Error with the configuration file")
             self.__logger.exception(exc)
@@ -75,7 +76,9 @@ class Grader:
         if self.__is_skipping_venv_creation or len(venv_checks) == 0:
             return scores
 
-        with VirtualEnvironment(self.__project_root, self.__is_keeping_venv):
+        venv_config = self.__config.get("venv", {})
+
+        with VirtualEnvironment(self.__project_root, is_keeping_venv_after_run=self.__is_keeping_venv, **venv_config):
             scores += [self.__run_check(check) for check in venv_checks]
 
         self.__cleanup()
@@ -95,11 +98,12 @@ class Grader:
         except CheckError as error:
             self.__logger.error("Check failed: %s", error)
 
+            # TODO - Pass the information from the exception
             match check:
                 case ScoredCheck():
-                    check_result = ScoredCheckResult(check.name, 0, check.max_points)
+                    check_result = ScoredCheckResult(check.name, 0, "", str(error), check.max_points)
                 case NonScoredCheck():
-                    check_result = NonScoredCheckResult(check.name, False)
+                    check_result = NonScoredCheckResult(check.name, False, "", str(error))
                 case _:
                     raise TypeError(f"Unknown check type: {type(check)}") from error