+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # pygrader
 
+## 1.8.0
+
+- Check results now show useful info and error messages
+- The venv created by pygrader is now with a non-standard name
+- Deleting existing venv is now optional
+- If a zip contains only the project as a subdirectory, use the subdir instead
+- Requirements check can now check for project installation
+
 ## 1.7.1
 
 - Subpackages are now properly exported